+++ src/Command/Traits/FormatterTrait.php
@@ -18,7 +18,7 @@ use Arvodia\Grouper\Grouper;
  * Description
  * 
  * @name    : FormatterTrait
- * @see     : 
+ * @see     :
  * @todo    : 
  *
  * @author Sidi Said Redouane <sidisaidredouane@live.com>
@@ -31,11 +31,6 @@ trait FormatterTrait {
         }
         $groups = $group ? [$group => $grouper->getGroup($group)] : $grouper->getGroups();
         $installedRepo = $this->getComposer()->getRepositoryManager()->getLocalRepository();
-        $locker = $this->getComposer()->getLocker();
-        $lockData = $locker->getLockData();
-
-        $packages = array_merge($lockData['packages'], $lockData['packages-dev']);
-
         foreach ($groups as $group => $groupConfig) {
             $groupStatus = [];
             foreach ($groups[$group]['require'] = $groupConfig['require'] ?? [] as $package => $packageConfig) {
@@ -46,8 +41,8 @@ trait FormatterTrait {
                 ];
             }
             $groups[$group]['status'] = empty($groupStatus) || in_array('no installed', $groupStatus) ? 'disable' : 'enabled';
-            if (!is_null($grouper->getGroupEnabled($group)) && $grouper->getGroupEnabled($group) !== ('enabled' == $groups[$group]['status'] ? true : false)) {
-                $this->getIO()->alert(sprintf('group "%1$s" in grouper.json "%2$s" but it is "%3$s".', $group, $grouper->getGroupEnabled($group) ? 'enabled' : 'disable', $groups[$group]['status']));
+            if (!is_null($grouper->isGroupActivated($group)) && $grouper->isGroupActivated($group) !== ('enabled' == $groups[$group]['status'] ? true : false)) {
+                $this->getIO()->alert(sprintf('group "%1$s" in grouper.json "%2$s" but these dependencies are "%3$s", use "composer grouper:groups %1$s -vv" for more detail.', $group, $grouper->isGroupActivated($group) ? 'enabled' : 'disable', $groups[$group]['status']));
             }
             if ($enabled && 'disable' == $groups[$group]['status']) {
                 unset($groups[$group]);