+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 CHANGELOG
 =========
 
+1.0.4
+-----
+* Code reduction
+* Print grouper version
+* Adding Verbose and Very Verbose for the groups command
+* Adding create and delete action groups manager
+* Now uninstall tasks are performed during group deactivation or held during uninstall packages
+* Function improvement nameValidator()
+* Adding option Run group tasks and these packages for the task command
+* Adding the uninstall option for group tasks
+* Disable POST_PACKAGE_UNINSTALL Events
+* Add more functions for class Grouper and FileSystems
+* Adding an example file
+
 1.0.1
 -----
 * Adding the Task Parameter for Group and Package